+++ laidoff/src/nav.cpp
@@ -87,7 +87,7 @@ void* load_nav(const char* filename) {
     char* d = create_binary_from_file(filename, &size);
 
     dtStatus status;
-    status = nav->nav_mesh->init(reinterpret_cast<unsigned char*>(d), size, DT_TILE_FREE_DATA);
+    status = nav->nav_mesh->init(reinterpret_cast<unsigned char*>(d), static_cast<int>(size), DT_TILE_FREE_DATA);
     if (dtStatusFailed(status)) {
         LOGE("nav->nav_mesh->init() error");
     }
@@ -524,7 +524,7 @@ void nav_update(LWNAV* nav, float move_speed, float delta_time) {
                 nav->path_query_output_location[i], nav->path_query_output_orientation[i]);
             // Jump jump!
             const float nav_jump_height = 0.7f;
-            nav->path_query_output_location[i][2] += nav_jump_height * fabs(sinf(nav->path_query[i].path_t * 11));
+            nav->path_query_output_location[i][2] += nav_jump_height * fabsf(sinf(nav->path_query[i].path_t * 11));
         }
     }
 }
